requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
docopt==0.6.2
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-MySQLdb==0.2.0
Flask-WTF==0.14.2
get==0.0.21
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
passlib==1.7.1
pipreqs==0.4.9
post==0.0.13
public==0.0.46
query-string==0.0.12
redis==2.10.6
requests==2.18.4
setupfiles==0.0.50
urllib3==1.22
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
yarg==0.1.9
